bugs
I had  a huge back yesterday with the login and add to cart button.

had to rewrite the code of login 4 times becuase at first the loginand sign up where in the same page and using segments to separate  the 2. and for that it was causing a huge issue because the code wasnt reading if the user is logged in or not so it wasnt let me add items in the basket.

than after removing the register code, i made the two in a separate pages and had a small bug and didnt know what it was. after realising it was an issue because i was still using the same code for that i had for the register and that was still causing a problem.after making the code again from scratch, the add to basket started to work. the fixed function is that if teh user is logged in,they can add items to the. basket but if they arent,they will be taken directly to the login pagee.